Use proper mime-types in the file input’s accept attribute
So far, the file input's accept attribute contained the list of allowed suffixes. While this works in Chrome, it does not in other Browsers. The HTML standard requires a list of MIME types [1]. With these changes, proper MIME types can be set optionally. This extends the browser support to Safari and also works in some cases for Firefox (e.g. "image/*" is supported by Firefox, but "image/png" is not). Suffixes are still used as fallback when no accepted MIME types are set, so that the existing behaviour is guaranteed.

[1] http://www.w3.org/TR/html-markup/input.file.html#input.file.attrs.accept