Make File Upload usable in arbitrary Forms
The handling of asynchronous requests is refactored, such that they can be handled automatically when ilDragDropFileInputGUI's checkInput method is called.

ilDragDropFileInputGUI can be used together with other input items in a form. One can choose to submit only the files (this is the previously defined behaviour, but makes the other input fields unusable), to submit all other input fields together with a file, or to submit first all files, followed with the remaining form data.

ilDragDropFileInputGUI can display a list of previously uploaded files. The standard jQuery-Fileupload API is used [1]. These files are displayed like successfully uploaded files.

Furthermore, a new option allows deleting previously uploaded files. A POST request is fired in the background with the file id.

[1] https://github.com/blueimp/jQuery-File-Upload/wiki/Setup#using-jquery-file-upload-ui-version-with-a-custom-server-side-upload-handler